Conflicts:
	pom.xml
	src/main/java/com/alibaba/middleware/benchmark/Test.java
	src/main/java/com/alibaba/middleware/conf/RaceConfig.java
	src/main/java/com/alibaba/middleware/handlefile/BuyerHandler.java
	src/main/java/com/alibaba/middleware/handlefile/ConstructSystem.java
	src/main/java/com/alibaba/middleware/handlefile/GoodHandler.java
	src/main/java/com/alibaba/middleware/race/OrderSystemImpl.java

+++ src/main/java/com/alibaba/middleware/handlefile/FileIndexWithOffset.java
@@ -7,10 +7,10 @@ package com.alibaba.middleware.handlefile;
  */
 public class FileIndexWithOffset {
 
-	public short fileIndex = 0;
+	public byte fileIndex = 0;
 	public long offset = 0;
 	
-	public FileIndexWithOffset( short fileIndex, long offset) {
+	public FileIndexWithOffset( byte fileIndex, long offset) {
 		this.fileIndex = fileIndex;
 		this.offset = offset;
 	}
